admin final

+++ admin 3/src/app/manageproviders/manageproviders.component.ts
@@ -1,5 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { Router } from '@angular/router';
+// import { DecimalPipe } from '@angular/common';
+// import { FormControl } from '@angular/forms';
 
 import {
   collection,
@@ -12,23 +14,50 @@ import {
 } from '@angular/fire/firestore';
 import { where,query } from '@firebase/firestore';
 import { AuthService } from '../auth.service';
+import{MatTableDataSource}from '@angular/material/table';
+import{MatPaginator}from '@angular/material/paginator';
+import{MatSort}from '@angular/material/sort';
 @Component({
   selector: 'app-manageproviders',
   templateUrl: './manageproviders.component.html',
   styleUrls: ['./manageproviders.component.css']
 })
-export class ManageprovidersComponent implements OnInit {
+export class ManageprovidersComponent implements OnInit{
   customers:any
   userData:any
+  //filter = new FormControl('');
+  displayedColumns=['name','email','phone','Delete']
+  dataSource!:MatTableDataSource<any>
+  @ViewChild('paginator') paginator! :MatPaginator;
+  @ViewChild(MatSort) matSort! :MatSort;
   constructor(private router : Router,private db2: Firestore,private authService:AuthService) { 
-    const Collection =query(collection(db2, 'users'),where("phoneNumber","!=",null)) ;
+    // var name=document.getElementById("selectname");
+    // var selected=name?.textContent;
+    const Collection =query(collection(db2, 'users'),where("phone","!=",null)) ;
     this.customers = collectionData(Collection);
     
+    collectionData(Collection).subscribe((data:any) => {
+      this.customers=data;
+      this.dataSource=new MatTableDataSource(this.customers);
+      this.dataSource.paginator=this.paginator;
+    this.dataSource.sort=this.matSort;
+    })
+    
+    
   }
-  
-
+  // search({ text, pipe }: { text: string; pipe: PipeTransform; }) {
+  //   return this.customers.filter((customer: { name: string; area: any; population: any; }) => {
+  //     const term = text.toLowerCase();
+  //     return customer.name.toLowerCase().includes(term)
+  //         || pipe.transform(customer.area).includes(term)
+  //         || pipe.transform(customer.population).includes(term);
+  //   });
+  // }
   ngOnInit(): void {
     this.userData=this.getUserInfo()
+  }
+  filterData(e:any){
+    this.dataSource.filter=e.target.value;
   }
  onclick(){
   this.router.navigate([`admin-dashboard/manageuser`]);
@@ -40,7 +69,10 @@ export class ManageprovidersComponent implements OnInit {
   this.router.navigate([`admin-dashboard/chart`]);
  }
  delete(id:any){
-  deleteDoc(doc(this.db2 , 'users' , id));
+  if (confirm("Are you sure you want to delete")) {
+    deleteDoc(doc(this.db2 , 'users' , id));
+  }
+  
  }
  getUserInfo(){
   return this.authService.getuser()
