+++ admin 3/src/app/manageusers/manageusers.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit,ViewChild } from '@angular/core';
 import { Router } from '@angular/router';
 import {
   Firestore,
@@ -11,7 +11,9 @@ import {
 } from '@angular/fire/firestore';
 import { where,query } from '@firebase/firestore';
 import { AuthService } from '../auth.service';
-
+import{MatTableDataSource}from '@angular/material/table';
+import{MatPaginator}from '@angular/material/paginator';
+import{MatSort}from '@angular/material/sort';
 @Component({
   selector: 'app-manageusers',
   templateUrl: './manageusers.component.html',
@@ -21,20 +23,28 @@ export class ManageusersComponent implements OnInit {
  //users:any
  providers:any
  userData:any
+ displayedColumns=['name','email','Delete','profile']
+ dataSource!:MatTableDataSource<any>
+ @ViewChild('paginator') paginator! :MatPaginator;
+ @ViewChild(MatSort) matSort! :MatSort;
 //  customers:any
   constructor(private router : Router,private db: Firestore,private authService:AuthService) {
     // const dataCollection = collection(db, 'users');
     // this.users = collectionData(dataCollection);
     const dataCollection =query(collection(db, 'users'),where("cr","!=",null)) ;
     this.providers = collectionData(dataCollection);
-    // collectionData(dataCollection).subscribe((data) => {
+    collectionData(dataCollection).subscribe((data) => {
+      this.providers=data;
+      this.dataSource=new MatTableDataSource(this.providers);
+      this.dataSource.paginator=this.paginator;
+    this.dataSource.sort=this.matSort;
     //   let j=-1;
     //   for(let i in data){
     //     if( data[i].cr){
     //       this.providers[j] =data[i];}
     //       // console.log(this.providers[j]) ;
     //       j++;
-    //   }
+      })
      
 
   //   // });
@@ -42,7 +52,9 @@ export class ManageusersComponent implements OnInit {
   // const Collection =query(collection(db, 'users'),where("phoneNumber","!=",null)) ;
   // this.customers = collectionData(Collection);
    }
-
+   filterData(e:any){
+    this.dataSource.filter=e.target.value;
+  }
   ngOnInit(): void {
     this.userData=this.getUserInfo()
   }
@@ -56,7 +68,10 @@ onclick2(){
   this.router.navigate([`admin-dashboard/managecustomers`]);
 }
 deleteMovie(id:any){
-  deleteDoc(doc(this.db , 'users' , id));
+  if (confirm("Are you sure you want to delete!")) {
+    deleteDoc(doc(this.db , 'users' , id));
+  }
+ 
 }
 getUserInfo(){
   return this.authService.getuser()
